Setting the time in a Date object in JavaScript
Setting the time in a Date object in JavaScript
When you and I created an object with a date like this: new Date(), we left the parentheses empty. In this case, we got the current time.

However, you can do it differently - pass it parameters in the format new Date (year, month, day, hours, minutes, seconds, milliseconds) and in this case, not the current time, but the one that we specified in the parameters. In this case, when passing as a parameter, the countdown of months starts from zero.

Parameters can be omitted from the end. In this case, the missing parameters for milliseconds, seconds and hours are considered equal to zero, and for days - one. Year and month cannot be omitted.

Let's create an object with a date for a specific point in time: